+++ lib/src/headers.dart
@@ -11,9 +11,13 @@ final _emptyHeaders = Headers._empty();
 
 /// Unmodifiable, key-insensitive header map.
 class Headers extends UnmodifiableMapView<String, List<String>> {
-  Map<String, String> _singleValues;
+  late final Map<String, String> singleValues = UnmodifiableMapView(
+    CaseInsensitiveMap.from(
+      map((key, value) => MapEntry(key, joinHeaderValues(value)!)),
+    ),
+  );
 
-  factory Headers.from(Map<String, List<String>> values) {
+  factory Headers.from(Map<String, List<String>>? values) {
     if (values == null || values.isEmpty) {
       return _emptyHeaders;
     } else if (values is Headers) {
@@ -25,14 +29,10 @@ class Headers extends UnmodifiableMapView<String, List<String>> {
 
   Headers._(Map<String, List<String>> values)
       : super(CaseInsensitiveMap.from(Map.fromEntries(values.entries
-            .where((e) => e.value?.isNotEmpty ?? false)
+            .where((e) => e.value.isNotEmpty)
             .map((e) => MapEntry(e.key, List.unmodifiable(e.value))))));
 
   Headers._empty() : super(const {});
-  factory Headers.empty() => _emptyHeaders;
 
-  Map<String, String> get singleValues => _singleValues ??= UnmodifiableMapView(
-        CaseInsensitiveMap.from(
-            map((key, value) => MapEntry(key, joinHeaderValues(value)))),
-      );
+  factory Headers.empty() => _emptyHeaders;
 }